+++ pkg/apis/ml/v1alpha1/extension_conditions.go
@@ -23,6 +23,7 @@ package v1alpha1
 import api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 
 const (
+	ExtensionStorageFoundCondition         api.ConditionType = "StorageFound"
 	ExtensionDeploymentFoundCondition      api.ConditionType = "DeploymentFound"
 	ExtensionMetadataServiceValidCondition api.ConditionType = "MetadataServiceValid"
 	LicenseValidCondition                  api.ConditionType = "LicenseValid"