+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - (Feature) (ML) Storage S3 sidecar implementation
 - (Feature) TLS CA Secret Key
 - (Refactoring) Extract Resource Helpers
+- (Feature) (ML) Extension Storage Condition
 
 ## [1.2.35](https://github.com/arangodb/kube-arangodb/tree/1.2.35) (2023-11-06)
 - (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks